Move DemoApplication
